+++ lab5/trapezoid_method.h
@@ -8,10 +8,10 @@ namespace trapezoid {
     double do_mid_calculations(
             double (&fun)(const double&),
             const double& a, const double& b,
-            const double& n);
+            const int& n);
 
     double solve_integral(
             double (&fun)(const double&),
           const double& a, const double& b,
-          const int& eps);
+          const double& eps);
 }
